refactor: arragenged everything in packages
